kafka_pub_sub/consumer/requirements.txt
kafka-python==2.0.2
PyYAML==6.0.1 